+++ dev/NetMath.js
@@ -125,6 +125,21 @@ class NetMath {
         return [...new Array(size)].map(v => Math.random()*2*limit-limit)
     }
 
+    static gaussian (size, {mean, stdDeviation}) {
+        return [...new Array(size)].map(() => {
+
+            let x1, x2, r, y
+
+            do {
+                x1 = 2 * Math.random() -1
+                x2 = 2 * Math.random() -1
+                r = x1**2 + x2**2
+            } while (r >= 1 || !r)
+
+            return mean + (x1 * (Math.sqrt(-2 * Math.log(r) / r))) * stdDeviation
+        })
+    }
+
     // Other
     static softmax (values) {
         const total = values.reduce((prev, curr) => prev+curr, 0)
@@ -133,6 +148,12 @@ class NetMath {
 
     static sech (value) {
         return (2*Math.exp(-value))/(1+Math.exp(-2*value))
+    }
+
+    static standardDeviation (arr) {
+        const avg = arr.reduce((p,c) => p+c) / arr.length
+        const diffs = arr.map(v => v - avg).map(v => v**2)
+        return Math.sqrt(diffs.reduce((p,c) => p+c) / diffs.length)
     }
 
     static maxNorm () {