+++ CHANGELOG.md
@@ -3,7 +3,13 @@
 #### Network
 - Reworked current weights initialization to be configurable
 - Set old weights init to uniform distribution, with additional limit config
+- Added mean as weightsConfig option, for gaussian distribution
+- Added stdDeviation as weightsConfig option, for gaussian distribution
 
+#### NetMath
+- Added standardDeviation to NetMath
+- Added gaussian weights distribution
+  
 # v1.3.0 - Regularization
 ---
 #### Network